Code framework optimization
