+++ src/battle_game/core/effects/regeneration.cpp
@@ -5,13 +5,13 @@
 namespace battle_game {
 
 std::string Regeneration::Description() const {
-  return (std::string) "持续回复生命值（每秒" +
-         std::to_string(health_per_second_) + (std::string) "，剩余" +
+  return (std::string) u8"持续回复生命值（每秒" +
+         std::to_string(health_per_second_) + (std::string) u8"，剩余" +
          std::to_string((float)tick_remain_ / kTickPerSecond) +
-         (std::string) "）";
+         (std::string) u8"）";
 }
 
-void Regeneration::Influence(Unit::Status &status) {
+void Regeneration::Influence(Status &status) {
   status.HealthChange(src_unit_id_, health_per_second_ / kTickPerSecond);
 }
