+++ src/battle_game/core/effects/speed_regulation.h
@@ -10,12 +10,12 @@ class SpeedRegulation : public Effect {
 
  public:
   std::string Name() const {
-    return "速度调节";
+    return u8"速度调节";
   }
   std::string Description() const {
-    return "单位速度受到与光标距离调节";
+    return u8"单位速度受到与光标距离调节";
   }
-  void Influence(Unit::Status &status);
+  void Influence(Status &status);
   uint32_t TickRemain() const;
   void TickPass();
   bool ShouldRemove() const;
